fix: allow empty .env file
Adjusts the environment variable validation to
allow the `.env` file to be empty so env variables
can be loaded from the environment instead of a
file on disk. This is required to deploy the
application in a kubernetes environment. This also
adds an empty `.env` file to the resulting image
so it won't error that the file can't be found.